fix: collectionNft called before intialized
