+++ src/nft/components/collection/CollectionAssetLoading.tsx
@@ -1,3 +1,4 @@
+import { ASSET_PAGE_SIZE } from 'graphql/data/nft/Asset'
 import { loadingAsset } from 'nft/css/loading.css'
 
 import SizingImage from '../../../assets/images/sizingImage.png'
@@ -5,7 +6,7 @@ import { Box } from '../../components/Box'
 import { Row } from '../Flex'
 import * as styles from './CollectionAssetLoading.css'
 
-export const CollectionAssetLoading = ({ height }: { height?: number }) => {
+const CollectionAssetLoading = ({ height }: { height?: number }) => {
   return (
     <Box as="div" className={styles.collectionAssetLoading}>
       <Box as="div" position="relative" width="full" style={{ height }}>
@@ -21,3 +22,11 @@ export const CollectionAssetLoading = ({ height }: { height?: number }) => {
     </Box>
   )
 }
+
+export const LoadingAssets = ({ count, height }: { count?: number; height?: number }) => (
+  <>
+    {Array.from(Array(count ?? ASSET_PAGE_SIZE), (_, index) => (
+      <CollectionAssetLoading key={index} height={height} />
+    ))}
+  </>
+)
