+++ src/nft/components/collection/CollectionNfts.tsx
@@ -55,8 +55,9 @@ import { useLocation } from 'react-router-dom'
 import styled, { css } from 'styled-components/macro'
 import { ThemedText } from 'theme'
 
-import { CollectionAssetLoading } from './CollectionAssetLoading'
+import { LoadingAssets } from './CollectionAssetLoading'
 import { MARKETPLACE_ITEMS, MarketplaceLogo } from './MarketplaceSelect'
+import { ClearAllButton } from './shared'
 import { Sweep } from './Sweep'
 import { TraitChip } from './TraitChip'
 
@@ -120,17 +121,6 @@ const ViewFullCollection = styled.span`
   ${OpacityHoverState}
 `
 
-export const ClearAllButton = styled.button`
-  color: ${({ theme }) => theme.textTertiary};
-  padding-left: 8px;
-  padding-right: 8px;
-  font-size: 14px;
-  font-weight: 600;
-  border: none;
-  cursor: pointer;
-  background: none;
-`
-
 const InfiniteScrollWrapper = styled.div`
   ${InfiniteScrollWrapperCss}
 `
@@ -171,14 +161,6 @@ const SweepText = styled(ThemedText.BodyPrimary)`
 const MarketNameWrapper = styled(Row)`
   gap: 8px;
 `
-
-export const LoadingAssets = ({ count, height }: { count?: number; height?: number }) => (
-  <>
-    {Array.from(Array(count ?? ASSET_PAGE_SIZE), (_, index) => (
-      <CollectionAssetLoading key={index} height={height} />
-    ))}
-  </>
-)
 
 const CollectionNftsLoading = ({ height }: { height?: number }) => (
   <Box width="full" className={styles.assetList}>
