+++ src/nft/components/collection/shared.tsx
@@ -1,0 +1,12 @@
+import styled from 'styled-components/macro'
+
+export const ClearAllButton = styled.button`
+  color: ${({ theme }) => theme.textTertiary};
+  padding-left: 8px;
+  padding-right: 8px;
+  font-size: 14px;
+  font-weight: 600;
+  border: none;
+  cursor: pointer;
+  background: none;
+`
